refactor: brakes upload many images into component as it might improve rerender performance
