+++ src/app/images/collections/[id]/CollectionAdmin.tsx
@@ -1,11 +1,8 @@
 'use client'
 import styles from './CollectionAdmin.module.scss'
-import { createImagesAction } from '@/actions/images/create'
 import { updateImageCollectionAction } from '@/actions/images/collections/update'
 import Form from '@/components/Form/Form'
 import TextInput from '@/components/UI/TextInput'
-import Dropzone from '@/components/UI/Dropzone'
-import PopUp from '@/components/PopUp/PopUp'
 import { destroyImageCollectionAction } from '@/actions/images/collections/destroy'
 import { ImageSelectionContext } from '@/contexts/ImageSelection'
 import { ImagePagingContext } from '@/contexts/paging/ImagePaging'
@@ -13,16 +10,13 @@ import Image from '@/components/Image/Image'
 import ImageUploader from '@/components/Image/ImageUploader'
 import useEditing from '@/hooks/useEditing'
 import VisibilityAdmin from '@/components/VisiblityAdmin/VisibilityAdmin'
-import { maxNumberOfImagesInOneBatch } from '@/services/images/ConfigVars'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faQuestion, faTrash, faUpload } from '@fortawesome/free-solid-svg-icons'
+import { faQuestion, faTrash } from '@fortawesome/free-solid-svg-icons'
 import { useRouter } from 'next/navigation'
-import { useContext, useState } from 'react'
-import { v4 as uuid } from 'uuid'
-import type { FileWithStatus } from '@/components/UI/Dropzone'
+import { useContext } from 'react'
 import type { VisibilityCollapsed } from '@/services/visibility/Types'
 import type { ExpandedImageCollection } from '@/services/images/collections/Types'
-import type { ActionReturn } from '@/actions/Types'
+import CollectionAdminUpload from './CollectionAdminUpload'
 
 type PropTypes = {
     collection: ExpandedImageCollection
@@ -31,7 +25,6 @@ type PropTypes = {
 
 export default function CollectionAdmin({ collection, visibility }: PropTypes) {
     const { id: collectionId, coverImage } = collection
-    const [files, setFiles] = useState<FileWithStatus[]>([])
     const router = useRouter()
     const selection = useContext(ImageSelectionContext)
     const pagingContext = useContext(ImagePagingContext)
@@ -49,70 +42,12 @@ export default function CollectionAdmin({ collection, visibility }: PropTypes) {
         }
     }
 
-    const handleBatchedUpload = async () => {
-        // split files into batches of maxNumberOfImagesInOneBatch
-        const batches = files.reduce((acc, file, index) => {
-            if (index % maxNumberOfImagesInOneBatch === 0) {
-                acc.push([])
-            }
-            acc[acc.length - 1].push(file)
-            return acc
-        }, [] as FileWithStatus[][])
-
-        let res: ActionReturn<void> = { success: true, data: undefined }
-        for (const batch of batches) {
-            const formData = new FormData()
-            batch.forEach(file => {
-                formData.append('files', file)
-            })
-            setFiles(prev => prev.map(file => {
-                if (batch.includes(file)) {
-                    return { ...file, uploadStatus: 'uploading' }
-                }
-                return file
-            }))
-            res = await createImagesAction.bind(null, collectionId)(formData)
-            if (res.success) {
-                setFiles(prev => prev.map(file => {
-                    if (batch.includes(file)) {
-                        return { ...file, uploadStatus: 'done' }
-                    }
-                    return file
-                }))
-            } else {
-                setFiles(prev => prev.map(file => {
-                    if (batch.includes(file)) {
-                        return { ...file, uploadStatus: 'error' }
-                    }
-                    return file
-                }))
-                return res
-            }
-        }
-        return res
-    }
-
     return (
         <>
             <aside className={styles.CollectionAdmin}>
                 <div className={styles.upload}>
                     <ImageUploader collectionId={collectionId} successCallback={refreshImages} />
-                    <PopUp PopUpKey={uuid()} showButtonContent={
-                        <>
-                        Last opp mange
-                            <FontAwesomeIcon icon={faUpload} />
-                        </>
-                    }>
-                        <Form
-                            className={styles.uploadMany}
-                            successCallback={refreshImages}
-                            title="last opp bilder"
-                            submitText="last opp"
-                            action={handleBatchedUpload}
-                        >
-                            <Dropzone label="last opp" name="files" files={files} setFiles={setFiles}/>
-                        </Form>
-                    </PopUp>
+                    <CollectionAdminUpload collectionId={collectionId} refreshImages={refreshImages} />
                 </div>
                 <Form
                     successCallback={() => {
